Changes based on PR comments: - removed airflow xcom push file , moved to decorator code - removed prefix configuration - nit fixes.
