+++ metaflow/metaflow_config.py
@@ -197,10 +197,6 @@ KUBERNETES_CONTAINER_REGISTRY = (
 )
 #
 
-###
-# Airflow Configuration
-###
-AIRFLOW_STATE_MACHINE_PREFIX = from_conf("METAFLOW_AIRFLOW_STATE_MACHINE_PREFIX")
 
 ###
 # Conda configuration
